+++ src/Credfeto.DotNet.Code.Analysis.Overrides/Ini/Section.cs
@@ -12,14 +12,14 @@ namespace Credfeto.DotNet.Code.Analysis.Overrides.Ini;
 [DebuggerDisplay("{Name} Order {Order}")]
 internal sealed class Section : INamedSection
 {
-    private readonly Dictionary<string, PropertyValue> _properties;
+    private readonly OrderedDictionary<string, PropertyValue> _properties;
     private readonly ISettings _settings;
     private List<string> _sectionComments;
 
     public Section(ISettings settings, int order, string? name, IReadOnlyList<string> sectionComments)
     {
         this._settings = settings;
-        this._sectionComments = [..Comments.Clean(sectionComments)];
+        this._sectionComments = [.. Comments.Clean(sectionComments)];
         this.Order = order;
         this.Name = name;
         this._properties = new(StringComparer.OrdinalIgnoreCase);
@@ -34,9 +34,7 @@ internal sealed class Section : INamedSection
 
     public string? Get(string key)
     {
-        return this._properties.TryGetValue(key: key, out PropertyValue? propertyValue)
-            ? propertyValue.Value
-            : null;
+        return this._properties.TryGetValue(key: key, out PropertyValue? propertyValue) ? propertyValue.Value : null;
     }
 
     public void Set(string key, string value)
@@ -61,10 +59,7 @@ internal sealed class Section : INamedSection
     {
         if (this._properties.TryGetValue(key: key, out PropertyValue? propertyValue))
         {
-            propertyValue.Comments =
-            [
-                ..comments.Select(Comments.Parse)
-            ];
+            propertyValue.Comments = [.. comments.Select(Comments.Parse)];
 
             return;
         }
@@ -98,15 +93,12 @@ internal sealed class Section : INamedSection
 
     public IReadOnlyList<string> SectionComment()
     {
-        return [..this._sectionComments];
+        return [.. this._sectionComments];
     }
 
     public void SectionComment(IReadOnlyList<string> comments)
     {
-        this._sectionComments =
-        [
-            ..comments.Select(Comments.Parse)
-        ];
+        this._sectionComments = [.. comments.Select(Comments.Parse)];
     }
 
     public ISettings ToSettings()
@@ -118,14 +110,14 @@ internal sealed class Section : INamedSection
     {
         if (!string.IsNullOrWhiteSpace(this.Name))
         {
-            stringBuilder = stringBuilder.AppendComments(comments: this._sectionComments)
-                                         .AppendLine($"[{this.Name}]");
+            stringBuilder = stringBuilder.AppendComments(comments: this._sectionComments).AppendLine($"[{this.Name}]");
         }
 
         foreach ((string key, PropertyValue propertyValue) in this._properties)
         {
-            stringBuilder = stringBuilder.AppendComments(comments: propertyValue.Comments)
-                                         .AppendProperty(key: key, value: propertyValue.Value, propertyValue.LineComment.Parse());
+            stringBuilder = stringBuilder
+                .AppendComments(comments: propertyValue.Comments)
+                .AppendProperty(key: key, value: propertyValue.Value, propertyValue.LineComment.Parse());
         }
 
         return stringBuilder;